+++ src/archive_text.c
@@ -50,7 +50,7 @@ GEM_INLINE archive_text_t* archive_text_read_mem(mm_t* const memory_manager) {
   // Read Header
   const uint64_t archive_text_model_no = mm_read_uint64(memory_manager);
   gem_cond_fatal_error(archive_text_model_no!=ARCHIVE_TEXT_MODEL_NO,
-      ARCHIVE_TEXT_WRONG_MODEL_NO,archive_text_model_no,ARCHIVE_TEXT_MODEL_NO);
+      ARCHIVE_TEXT_WRONG_MODEL_NO,archive_text_model_no,(uint64_t)ARCHIVE_TEXT_MODEL_NO);
   archive_text->hypertext = (mm_read_uint64(memory_manager)==1); // Hypertext
   archive_text->run_length = (mm_read_uint64(memory_manager)==1); // RL-Text
   archive_text->explicit_complement = (mm_read_uint64(memory_manager)==1); // Explicit RC-text
